Object Documentation

sensemakr object

Sensitivity analysis to unobserved confounders.

The sensemakr type performs sensitivity analysis to omitted variables as discussed in Cinelli and Hazlett (2020). It is comprized of several sensitivity statistics for reporting. After creating the object, you may directly use the plot and summary methods of the returned object.

sensemakr is a convenience type. You may use the other sensitivity functions of the package directly, such as the functions for sensitivity plots (ovb_contour_plot, ovb_extreme_plot), the functions for computing bias-adjusted estimates and t-values (adjusted_estimate, adjusted_t), the functions for computing the robustness value and partial $R^2$ (robustness_value, partial_r2), or the functions for bounding the strength of unobserved confounders (ovb_bounds), among others.

Sensemakr.sensemakr — Type

Sensemakr.sensemakr

Object comprised of necessary parameters and statistics for sensitivity analysis

Base.summary — Method

Base.summary(sense_obj::sensemakr, digits::Int64 = 3)

Print a summary of the sensitivity results for sense_obj, including robustness value, extreme confounding scenario, and benchmark bounding.

Arguments:

sense_obj: object to be summarized.

digits (default: 3): number of digits to round numbers to.

Sensemakr.plot — Function

Sensemakr.plot(sense_obj::sensemakr; kwargs...)

Provide the contour and extreme scenario sensitivity plots of the sensitivity analysis results obtained with the function sensemakr

They are basically dispatchers to the core plot functions ovbcontourplot and ovbextremeplot

Arguments:

sense_obj: object with the statistics to be plotted.

plot_type (default: "contour"): Either "extreme" or "contour".

sensitivity_of (default: "estimate"): Either "estimate" or "t-value"

Sensemakr.ovb_minimal_reporting — Function

ovb_minimal_reporting(sense_obj::sensemakr, digits::Int64 = 3, res_display::Bool = true)

This function returns the HTML code for a table summarizing the sensemakr object.

Arguments:

digits (default: 3): number of digits to round numbers to.

res_display (default: true): whether to display the table.

Plotting functions

Sensemakr provides functions for creating sensitivity contour plots and extreme scenario sensitivity plots. They can be used on an object of class sensemakr, directly with an OLS StatsModels.TableRegressionModel object, or by providing the required statistics manually.

Sensemakr.ovb_contour_plot — Function

ovb_contour_plot(estimate::Float64, se::Float64, dof::Int64, kwargs...)

Contour plots of omitted variable bias for sensitivity analysis.

The horizontal axis of the plot shows hypothetical values of the partial R2 of the unobserved confounder(s) with the treatment. The vertical axis shows hypothetical values of the partial R2 of the unobserved confounder(s) with the outcome. The contour levels represent the adjusted estimates (or t-values) of the treatment effect.

The reference points are the bounds on the partial R2 of the unobserved confounder if it were k times “as strong” as the observed covariates used for benchmarking (see arguments kd and ky). The dotted red line show the chosen critical threshold (for instance, zero): confounders with such strength (or stronger) are sufficient to invalidate the research conclusions. All results are exact for single confounders and conservative for multiple/nonlinear confounders.

See Cinelli and Hazlett (2020) for details.

Arguments:

estimate: estimated parameter of the variable of interest

se: standard error of the estimated parameter of the variable of interest

dof degrees of freedom of the estimated regression model

r2dz_x (default: nothing): number or array containing hypothetical partial R2 of unobserved confounder Z with treatment D, given covariates X.

r2yz_dx (default: nothing): number or array containing hypothetical partial R2 of unobserved confounder Z with outcome Y, given covariates X and treatment D.

sensitivity_of (default: "estimate"): either "estimate" or "t-value"

kd (default: 1): a float or vector of floats with each being a multiple of the strength of association between a benchmark variable and the treatment variable to test with benchmark bounding.

ky (default: nothing): a float or vector of floats with each being a multiple of the strength of association between a benchmark variable and the treatment variable to test with benchmark bounding.

benchmark_covariates (default: nothing): a string or vector of strings with the names of the variables to use for benchmarking.

bound_label (default: nothing): label of the bound variable.

reduce (default: true): whether to reduce (true) or increase (false) the estimate due to putative confounding.

estimate_threshold (default: 0): threshold line to emphasize when contours correspond to estimate.

t_threshold (default: 2): threshold line to emphasize when contours correspond to t-value.

lim (default: nothing): x axis maximum.

lim_y (default: nothing): y axis maximum.

col_contour (default: "black"): color of contour lines.

col_thr_line (default: "red"): color of the threshold contour line.

label_text (default: true): whether to include label text.

label_bump_x (default: nothing): x axis position of label above 0.

label_bump_y (default: nothing): y axis position of label above 0.

plot_margin_fraction (default: 0.05): margin fraction added to the top of lim and lim_y.

round_dig (default: 3): rounding digit of the display numbers.

n_levels (default: nothing): maximum number of countours in the plot.

ovb_contour_plot(model::StatsModels.TableRegressionModel, treatment::String, kwargs...)

Uses model and treatment to obtain the estimate, standard error, and degrees of freedom; it then dispatches them to the relevant method.

ovb_contour_plot(sense_obj::sensemakr, kwargs...)

Extracts the estimate, standard error, and degrees of freedom from sense_obj; it then dispatches them to the relevant method.

Sensemakr.add_bound_to_contour — Function

add_bound_to_contour(r2dz_x::Union{Real, Array{<:Real}}, r2yz_dx::Union{Real, Array{<:Real}; kwargs...}

Add bound label to the contour plot of omitted variable bias for sensitivity analysis.

The reference points are the bounds on the partial R2 of the unobserved confounder if it were k times "as strong" as the observed covariate used for benchmarking (see arguments kd and ky).

Arguments

r2dz_x: partial $R^2$ of a putative unobserved confounder "z" with the treatment variable "d", with observed covariates "x" partialed out, as implied by z being kd-times as strong as the benchmark_covariates.

r2yz_dx partial $R^2$ of a putative unobserved confounder "z" with the outcome variable "y", with observed covariates "x" and the treatment variable "d" partialed out, as implied by z being ky-times as strong as the benchmark_covariates.

kd (default: 1): a float or vector of floats with each being a multiple of the strength of association between a benchmark variable and the treatment variable to test with benchmark bounding.

ky (default: nothing): same as kd except measured in terms of strength of association with the outcome variable.

bound_value (default: nothing): the value of the reference point.

bound_label (default: nothing): a string that label the reference point.

sensitivity_of (default: nothing): either “estimate” or “t-value”.

label_text (default: true): whether to include the label text.

label_bump_x (default: nothing): x axis position of label above 0.

label_bump_y (default: nothing): y axis position of label above 0.

n_levels (default: nothing): maximum number of countours in the plot.

add_bound_to_contour(bound::DataFrame; kwargs...)

Extracts r2dzx, r2yzdx, and bound_label vectors from bound and dispatches them to the relevant method.

Sensemakr.ovb_extreme_plot — Function

ovb_extreme_plot(estimate::Float64, se::Float64, dof::Int64; kwargs...)

Extreme scenario plots of omitted variable bias for sensitivity analysis.

The horizontal axis shows the partial R2 of the unobserved confounder(s) with the treatment. The vertical axis shows the adjusted treatment effect estimate. The partial R2 of the confounder with the outcome is represented by different curves for each scenario, as given by the parameter r2yz_dx. The red marks on horizontal axis are bounds on the partial R2 of the unobserved confounder kd times as strong as the covariates used for benchmarking. The dotted red line represent the threshold for the effect estimate deemed to be problematic (for instance, zero).

See Cinelli and Hazlett (2020) for details.

Arguments:

estimate: estimated parameter of the variable of interest

se: standard error of the estimated parameter of the variable of interest

dof degrees of freedom of the estimated regression model

sensitivity_of (default: "estimate"): either "estimate" or "t-value". This function is currently only defined for "estimate"

benchmark_covariates (default: nothing): a string or vector of strings with the names of the variables to use for benchmarking.

kd (default: 1): a float or vector of floats with each being a multiple of the strength of association between a benchmark variable and the treatment variable to test with benchmark bounding.

ky (default: nothing): a float or vector of floats with each being a multiple of the strength of association between a benchmark variable and the treatment variable to test with benchmark bounding.

r2dz_x (default: nothing): a float or vector of floats with the partial $R^2$ of a putative unobserved confounder "z" with the treatment variable "d", with observed covariates "x" partialed out, as implied by z being kd-times as strong as the benchmark_covariates.

r2yz_dx (default: [1.0, 0.75, 0.5]): a float or vector of floats with the partial $R^2$ of a putative unobserved confounder "z" with the outcome variable "y", with observed covariates "x" and the treatment variable "d" partialed out, as implied by z being ky-times as strong as the benchmark_covariates.

reduce (default: true): whether to reduce (true) or increase (false) the estimate due to putative confounding.

threshold (default: 0): threshold line to emphasize when contours correspond to estimate.

lim (default: nothing): x axis maximum.

lim_y (default: nothing): y axis maximum.

xlab (default: nothing): x axis label text.

ylab (default: nothing): y axis label text.

ovb_extreme_plot(sense_obj::sensemakr; kwargs...)

Extracts the estimate, standard error, and degrees of freedom from sense_obj; it then dispatches them to the relevant method.

Sensitivity Bounds

Bounds on the strength of unobserved confounders using observed covariates, as in Cinelli and Hazlett (2020).

These functions may be useful when computing benchmarks for using only summary statistics from papers you see in print.

Currently only the bounds based on partial R2 are implemented. Other bounds will be implemented soon.

Sensemakr.ovb_bounds — Function

ovb_bounds(model::StatsModels.TableRegressionModel, treatment::String; kwargs...)

Returns a DataFrame which provides bounds on the strength of unobserved confounders using observed covariates, as in Cinelli and Hazlett (2020).

Arguments

model: object for the restricted regression you have provided.

treatment: name of the treatment variable or variable of interest.

benchmark_covariates: names of variables to use for benchmark bounding.

kd (default: 1): multiple of the strength of association between a benchmark variable and the treatment.

ky: same as kd except measured in terms of strength of association with the outcome variable.

alpha (default: 0.05): significance level for the robustness value $RV_{qa}$ to render the estimate not significant.

h0 (default: 0): null hypothesis effect size.

reduce (default: true): whether to reduce (true) or increase (false) the estimate due to putative computing.

bound (default: "partial_r2"): type of bound to perform; as of now, only partial $R^2$ bounding is allowed.

adjusted_estimates (default: true): whether to compute bias-adjusted estimates, standard errors, and t-statistics.

Sensemakr.ovb_partial_r2_bound — Function

ovb_partial_r2_bound(; kwargs...)

Returns a DataFrame with the bounds on the strength of the unobserved confounder.

Adjusted estimates, standard errors and t-values (among other quantities) need to be computed manually by the user using those bounds with the functions adjustedestimate, adjustedse and adjusted_t.

Arguments

model: StatsModels.TableRegressionModel object for the restricted regression you have provided.

treatment: string with the name of the treatment variable of interest.

r2dxj_x: float with the partial R2 of covariate Xj with the treatment D (after partialling out the effect of the remaining covariates X, excluding Xj).

r2yxj_dx: float with the partial R2 of covariate Xj with the outcome Y (after partialling out the effect of the remaining covariates X, excluding Xj).

benchmark_covariates: a string or vector of strings with names of the variables to use for benchmark bounding.

kd (default: 1): a float or list of floats with each being a multiple of the strength of association between a benchmark variable and the treatment variable to test with benchmark bounding.

ky (default: nothing): same as kd except measured in terms of strength of association with the outcome variable.

Sensitivity Statistics

Computes the sensitivity statistics: robustness value, partial R2, and Cohen’s f2.

Sensemakr.robustness_value — Function

robustness_value(; model::Union{Nothing, StatsModels.TableRegressionModel} = nothing, covariates::Union{Nothing, String, Vector{String}} = nothing, t_statistic::Union{Real, Nothing} = nothing, dof::Union{Nothing, Int64} = nothing, kwargs...)

Compute the robustness value of a regression coefficient.

The robustness value describes the minimum strength of association (parameterized in terms of partial R2) that omitted variables would need to have both with the treatment and with the outcome to change the estimated coefficient by a certain amount (for instance, to bring it down to zero).

For instance, a robustness value of 1% means that an unobserved confounder that explain 1% of the residual variance of the outcome and 1% of the residual variance of the treatment is strong enough to explain away the estimated effect. Whereas a robustness value of 90% means that any unobserved confounder that explain less than 90% of the residual variance of both the outcome and the treatment assignment cannot fully account for the observed effect. You may also compute robustness value taking into account sampling uncertainty. See details in Cinelli and Hazlett (2020).

Arguments:

model: object for the restricted regression you have provided.

covariates: names of the variables to use for benchmark bounding.

t_statistic: the t_statistic for the restricted model regression.

dof: degrees of freedom of the restricted regression

q (default: 1): a float with the percent to reduce the point estimate by for the robustness value RV_q.

alpha (default = 1.0): a float with the significance level for the robustness value RV_qa to render the estimate not significant.

Sensemakr.partial_r2 — Function

partial_r2(; model::Union{Nothing, StatsModels.TableRegressionModel} = nothing, covariates::Union{Nothing, String, Vector{String}} = nothing, t_statistic::Union{Real, Nothing} = nothing, dof::Union{Nothing, Int64} = nothing)

Compute the partial R2 for a linear regression model.

The partial R2 describes how much of the residual variance of the outcome (after partialing out the other covariates) a covariate explains.

The partial R2 can be used as an extreme-scenario sensitivity analysis to omitted variables. Considering an unobserved confounder that explains 100% of the residual variance of the outcome, the partial R2 describes how strongly associated with the treatment this unobserved confounder would need to be in order to explain away the estimated effect.

For details see Cinelli and Hazlett (2020).

Arguments:

model: object for the restricted regression you have provided.

covariates: names of the variables to use for benchmark bounding.

t_statistic: the t_statistic for the restricted model regression.

dof: degrees of freedom of the restricted regression

Sensemakr.partial_f2 — Function

partial_f2(; model::Union{Nothing, StatsModels.TableRegressionModel} = nothing, covariates::Union{Nothing, String, Vector{String}} = nothing, t_statistic::Union{Real, Nothing} = nothing, dof::Union{Nothing, Int64} = nothing)

Compute the partial (Cohen's) f2 for a linear regression model.

The partial (Cohen's) f2 is a common measure of effect size (a transformation of the partial R2) that can also be used directly for sensitivity analysis using a bias factor table. For details see Cinelli and Hazlett (2020).

Arguments

model: object for the restricted regression you have provided.

covariates: names of the variables to use for benchmark bounding.

t_statistic: the t_statistic for the restricted model regression.

dof: degrees of freedom of the restricted regression

Sensemakr.group_partial_r2 — Function

Partial R2 of groups of covariates in a linear regression model.

This function computes the partial R2 of a group of covariates in a linear regression model. Multivariate version of the partial_r2 function; see that for more details.

Arguments

model: object for the restricted regression you have provided.

covariates: names of the variables to use for benchmark bounding.

t_statistic: t_statistic for the restricted model regression.

p: number of parameters.

dof: degrees of freedom of the restricted regression.

Bias functions

Compute bias-adjusted estimates, standard-errors, and t-values.

All methods in the script below have similar purposes and parameters, so they are all described here.

These functions compute bias adjusted estimates (adjustedestimate), standard-errors (adjustedse), and t-values (adjusted_t), given a hypothetical strength of the confounder in the partial R2 parameterization.

They return a vector with the adjusted estimate, standard error, or t-value for each partial R2 passed in.

Internally, we also have functions defined to compute the bias and relative_bias, given the same arguments. We also define internal functions to compute the bias function and relative bias function for the partial R2 parameters.

Sensemakr.adjusted_estimate — Function

adjusted_estimate(r2dz_x::Union{Float64, Vector{Float64}}, r2yz_dx::Union{Float64, Vector{Float64}}; kwargs...)

Compute the bias-adjusted coefficient estimate.

Arguments

r2dz_x: partial $R^2$ of a putative unobserved confounder "z" with the treatment variable "d", with observed covariates "x" partialed out.

r2yz_dx: partial $R^2$ of a putative unobserved confounder "z" with the outcome "y", with observed covariates "x" and treatment "d" partialed out.

model: StatsModels.TableRegressionModel object for the restricted regression you have provided.

treatment: string with the name of the treatment variable of interest.

estimate: float with the unadjusted estimate of the coefficient for the independent variable of interest.

se: float with the unadjusted standard error of the regression.

dof: an int with the degrees of freedom of the regression.

reduce (default: true): whether to reduce (true) or increase (false) the estimate due to putative confounding.

Sensemakr.adjusted_se — Function

adjusted_se(r2dz_x::Union{Float64, Vector{Float64}}, r2yz_dx::Union{Float64, Vector{Float64}}; kwargs...)

Compute the bias-adjusted regression standard error.

Arguments

r2dz_x: partial $R^2$ of a putative unobserved confounder "z" with the treatment variable "d", with observed covariates "x" partialed out.

r2yz_dx: partial $R^2$ of a putative unobserved confounder "z" with the outcome "y", with observed covariates "x" and treatment "d" partialed out.

model: StatsModels.TableRegressionModel object for the restricted regression you have provided.

treatment: string with the name of the treatment variable of interest.

se: float with the unadjusted standard error of the regression.

dof: an int with the degrees of freedom of the regression.

Sensemakr.adjusted_t — Function

adjusted_t(r2dz_x::Union{Float64, Vector{Float64}}, r2yz_dx::Union{Float64, Vector{Float64}}; kwargs...)

Compute bias-adjusted t-statistic, (adjustedestimate - h0) / adjustedse.

Arguments

r2dz_x: partial $R^2$ of a putative unobserved confounder "z" with the treatment variable "d", with observed covariates "x" partialed out.

r2yz_dx: partial $R^2$ of a putative unobserved confounder "z" with the outcome "y", with observed covariates "x" and treatment "d" partialed out.

model: StatsModels.TableRegressionModel object for the restricted regression you have provided.

treatment: string with the name of the treatment variable of interest.

estimate: float with the unadjusted estimate of the coefficient for the independent variable of interest.

se: float with the unadjusted standard error of the regression.

dof: an int with the degrees of freedom of the regression.

reduce (default: true): whether to reduce (true) or increase (false) the estimate due to putative confounding.

h0: (default: 0): the test value for null hypothesis.

Sensemakr.adjusted_partial_r2 — Function

adjusted_partial_r2(r2dz_x::Union{Float64, Vector{Float64}}, r2yz_dx::Union{Float64, Vector{Float64}}; kwargs...)

Compute the bias-adjusted partial R2, based on adjusted_t

Arguments

r2dz_x: partial $R^2$ of a putative unobserved confounder "z" with the treatment variable "d", with observed covariates "x" partialed out.

r2yz_dx: partial $R^2$ of a putative unobserved confounder "z" with the outcome "y", with observed covariates "x" and treatment "d" partialed out.

model: StatsModels.TableRegressionModel object for the restricted regression you have provided.

treatment: string with the name of the treatment variable of interest.

estimate: float with the unadjusted estimate of the coefficient for the independent variable of interest.

se: float with the unadjusted standard error of the regression.

dof: an int with the degrees of freedom of the regression.

reduce (default: true): whether to reduce (true) or increase (false) the estimate due to putative confounding.

h0: (default: 0): the test value for null hypothesis.

Sensemakr.bias — Function

bias(r2dz_x::Union{Float64, Vector{Float64}}, r2yz_dx::Union{Float64, Vector{Float64}}; kwargs...)

Compute the omitted variable bias for the partial R2 parameterization.

Arguments

r2dz_x: partial $R^2$ of a putative unobserved confounder "z" with the treatment variable "d", with observed covariates "x" partialed out.

r2yz_dx: partial $R^2$ of a putative unobserved confounder "z" with the outcome "y", with observed covariates "x" and treatment "d" partialed out.

model: StatsModels.TableRegressionModel object for the restricted regression you have provided.

treatment: string with the name of the treatment variable of interest.

se: float with the unadjusted standard error of the regression.

dof: an int with the degrees of freedom of the regression.

#```@docs

Sensemakr.relative_bias

## Data

Provides the example data for the package.

@docs Sensemakr.load_darfur ```

This page was generated using Literate.jl.